requirements.txt
beautifulsoup4
git+https://github.com/Rapptz/discord.py.git
discord-ext-voice-recv
Pillow
pvporcupine
pvrhino
python-dotenv
pyttsx3
PyYAML
Requests
SpeechRecognition
yt_dlp
PyNaCl
numpy
resampy
Flask
Flask_SocketIO
watchdog
GitPython
validators
pytest
pytest-asyncio
pytest-mock
tzlocal
parsedatetime
tzdata